Add touch mode toggle to mobile titlebar
