+++ lib/views/app_title_bar/app_title_bar_desktop.dart
@@ -8,7 +8,6 @@ class _AppTitleBarDesktop extends StatelessWidget {
     bool canGoBack = context.select((AppModel m) => m.canPopNav);
     bool showBackBtn = isGuestUser == false && canGoBack;
     double appWidth = context.widthPx;
-
     bool isMac = UniversalPlatform.isMacOS;
     bool isMobile = context.widthPx < Sizes.smallPhone;
     bool showTouchToggle = true; //context.select((AppModel m) => m.hasUser) == true;
@@ -24,72 +23,18 @@ class _AppTitleBarDesktop extends StatelessWidget {
             HSpace(80), // Reserve some space for the native btns
             if (showBackBtn) _BackBtn(),
             Spacer(),
-            if (showTouchToggle) _TouchModeToggleBtn(invertPopupAlign: isMac),
+            if (showTouchToggle) TouchModeToggleBtn(invertPopupAlign: isMac),
             HSpace.sm,
             _AdaptiveProfileBtn(invertRow: true, useBottomSheet: isMobile),
           ] else ...[
             // Linux and Windows are left aligned and simple
             _AdaptiveProfileBtn(useBottomSheet: isMobile),
-            if (showTouchToggle) _TouchModeToggleBtn(invertPopupAlign: isMac),
+            if (showTouchToggle) TouchModeToggleBtn(invertPopupAlign: isMac),
             HSpace.sm,
             if (showBackBtn) _BackBtn(),
           ]
         ],
       ),
     ]);
-  }
-}
-
-class _TouchModeToggleBtn extends StatelessWidget {
-  const _TouchModeToggleBtn({Key key, @required this.invertPopupAlign}) : super(key: key);
-  final bool invertPopupAlign;
-
-  @override
-  Widget build(BuildContext context) {
-    void handleTouchModeToggled(bool value) {
-      ClosePopoverNotification().dispatch(context);
-      context.read<AppModel>().enableTouchMode = !value;
-    }
-
-    AppTheme theme = context.watch();
-    bool touchMode = context.select((AppModel m) => m.enableTouchMode);
-    double padding = touchMode ? 8 : 4;
-    return SimpleBtn(
-      onPressed: () => handleTouchModeToggled(touchMode),
-      child: Container(
-        width: 40,
-        child: Stack(
-          children: [
-            Positioned.fill(
-                child: TweenAnimationBuilder<double>(
-              tween: Tween(begin: padding, end: padding),
-              duration: Times.fast,
-              curve: Curves.easeOut,
-              builder: (_, animatedPadding, __) => Padding(
-                padding: EdgeInsets.all(animatedPadding),
-                child: Container(
-                  decoration: BoxDecoration(
-                    borderRadius: BorderRadius.circular(99),
-                    color: theme.grey.withOpacity(.3),
-                  ),
-                  child: PopOverRegion.hover(
-                    key: ValueKey(touchMode),
-                    popAnchor: invertPopupAlign ? Alignment.topRight : Alignment.topLeft,
-                    anchor: invertPopupAlign ? Alignment.bottomRight : Alignment.bottomLeft,
-                    popChild: StyledTooltip(touchMode ? "Switch to Precision Mode" : "Switch to Touch Mode",
-                        arrowAlignment: invertPopupAlign ? Alignment.topRight : Alignment.topLeft),
-                    child: MaterialIcon(
-                      touchMode ? Icons.mouse : Icons.fingerprint,
-                      color: theme.accent1,
-                      size: 22 - animatedPadding,
-                    ),
-                  ),
-                ),
-              ),
-            )),
-          ],
-        ),
-      ),
-    );
   }
 }
