+++ lib/views/app_title_bar/app_title_bar_mobile.dart
@@ -9,6 +9,7 @@ class _AppTitleBarMobile extends StatelessWidget {
         children: [
           if (showBackBtn) _BackBtn(),
           Spacer(),
+          TouchModeToggleBtn(invertPopupAlign: true),
           _AdaptiveProfileBtn(invertRow: true, useBottomSheet: true),
         ],
       ),
